+++ src/main/java/wiki/zimo/wiseduunifiedloginapi/entity/IapLoginEntity.java
@@ -1,26 +1,13 @@
 package wiki.zimo.wiseduunifiedloginapi.entity;
 
-public class IapLoginEntity {
-    private String loginUrl;
+public class IapLoginEntity extends BaseLoginEntity {
     private String doLoginUrl;
     private String itUrl;
-    private String needcaptchaUrl;
-    private String captchaUrl;
 
     public IapLoginEntity(String loginUrl, String doLoginUrl, String itUrl, String needcaptchaUrl, String captchaUrl) {
-        this.loginUrl = loginUrl;
+        super(loginUrl, needcaptchaUrl, captchaUrl);
         this.doLoginUrl = doLoginUrl;
         this.itUrl = itUrl;
-        this.needcaptchaUrl = needcaptchaUrl;
-        this.captchaUrl = captchaUrl;
-    }
-
-    public String getLoginUrl() {
-        return loginUrl;
-    }
-
-    public void setLoginUrl(String loginUrl) {
-        this.loginUrl = loginUrl;
     }
 
     public String getDoLoginUrl() {
@@ -39,30 +26,14 @@ public class IapLoginEntity {
         this.itUrl = itUrl;
     }
 
-    public String getNeedcaptchaUrl() {
-        return needcaptchaUrl;
-    }
-
-    public void setNeedcaptchaUrl(String needcaptchaUrl) {
-        this.needcaptchaUrl = needcaptchaUrl;
-    }
-
-    public String getCaptchaUrl() {
-        return captchaUrl;
-    }
-
-    public void setCaptchaUrl(String captchaUrl) {
-        this.captchaUrl = captchaUrl;
-    }
-
     @Override
     public String toString() {
         return "IapLoginEntity{" +
-                "loginUrl='" + loginUrl + '\'' +
-                ", doLoginUrl='" + doLoginUrl + '\'' +
-                ", itUrl='" + itUrl + '\'' +
-                ", needcaptchaUrl='" + needcaptchaUrl + '\'' +
-                ", captchaUrl='" + captchaUrl + '\'' +
+                "loginUrl='" + getLoginUrl() + '\'' +
+                ", doLoginUrl='" + getDoLoginUrl() + '\'' +
+                ", itUrl='" + getItUrl() + '\'' +
+                ", needcaptchaUrl='" + getNeedcaptchaUrl() + '\'' +
+                ", captchaUrl='" + getCaptchaUrl() + '\'' +
                 '}';
     }
 }
