+++ src/main/java/wiki/zimo/wiseduunifiedloginapi/process/impl/CuitCasLoginProcess.java
@@ -1,38 +1,26 @@
-package wiki.zimo.wiseduunifiedloginapi.process;
+package wiki.zimo.wiseduunifiedloginapi.process.impl;
 
-import net.sourceforge.tess4j.TesseractException;
 import org.jsoup.Connection;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 import wiki.zimo.wiseduunifiedloginapi.builder.CasLoginEntityBuilder;
-import wiki.zimo.wiseduunifiedloginapi.entity.CasLoginEntity;
-import wiki.zimo.wiseduunifiedloginapi.helper.AESHelper;
-import wiki.zimo.wiseduunifiedloginapi.helper.ImageHelper;
 import wiki.zimo.wiseduunifiedloginapi.helper.TesseractOCRHelper;
+import wiki.zimo.wiseduunifiedloginapi.process.OcrLoginProcess;
 import wiki.zimo.wiseduunifiedloginapi.trust.HttpsUrlValidator;
 
-import java.io.File;
-import java.io.IOException;
 import java.net.HttpURLConnection;
 import java.util.HashMap;
 import java.util.Map;
 
 /**
- * 武汉轻工大学认证
+ * 成都信息工程大学认证
  */
-public class WhpuCasLoginProcess {
-    private CasLoginEntity loginEntity;
-    private Map<String, String> params;
-
-    public WhpuCasLoginProcess(String loginUrl, Map<String, String> params) {
-        this.loginEntity = new CasLoginEntityBuilder()
-                .loginUrl(loginUrl)
-                .build();
-        loginEntity.setNeedcaptchaUrl("http://cas.whpu.edu.cn/authserver/checkNeedCaptcha.htl");
-        loginEntity.setCaptchaUrl("http://cas.whpu.edu.cn/authserver/getCaptcha.htl");
-        this.params = params;
+public class CuitCasLoginProcess extends OcrLoginProcess {
+    public CuitCasLoginProcess(String loginUrl, Map<String, String> params) {
+        super(loginUrl, params, CasLoginEntityBuilder.class);
+        loginEntity.setCaptchaUrl("https://sso.cuit.edu.cn/authserver/captcha");
     }
 
     public Map<String, String> login() throws Exception {
@@ -47,7 +35,7 @@ public class WhpuCasLoginProcess {
         headers.put("Accept-Language", "zh-CN,zh;q=0.9");
         headers.put("Cache-Control", "max-age=0");
         headers.put("Connection", "keep-alive");
-        headers.put("Host", "cas.whpu.edu.cn");
+        headers.put("Host", "authserver.cumt.edu.cn");
         headers.put("Upgrade-Insecure-Requests", "1");
         headers.put("User-Agent", "Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.125 Safari/537.36");
 
@@ -58,8 +46,9 @@ public class WhpuCasLoginProcess {
                 .headers(headers);
         Connection.Response res = con.execute();
 //        System.out.println(res.url());
-//        System.out.println(doc);
 
+        // 更新login_url
+        loginEntity.setLoginUrl(res.url().toString());
         // 全局cookie
         Map<String, String> cookies = res.cookies();
 
@@ -70,23 +59,14 @@ public class WhpuCasLoginProcess {
                 .cookies(cookies)
                 .get();
 
+//        System.out.println(doc);
+
         // 获取登陆表单
-        Element pwdLoginDiv = doc.getElementById("pwdLoginDiv");
-        Element form = pwdLoginDiv.getElementById("loginFromId");
+        Element form = doc.getElementById("fm1");
 
         if (form == null) {
-            throw new RuntimeException("网页中没有找到loginFromId，请联系开发者！！！");
+            throw new RuntimeException("网页中没有找到fm1，请联系开发者！！！");
         }
-
-        // 处理加密的盐
-        Element saltElement = doc.getElementById("pwdEncryptSalt");
-
-        String salt = null;
-        if (saltElement != null) {
-            salt = saltElement.val();
-        }
-
-//        System.out.println("盐是 " + salt);
 
         // 获取登陆表单里的输入
         Elements inputs = form.getElementsByTag("input");
@@ -109,24 +89,21 @@ public class WhpuCasLoginProcess {
         }
 
         params.put("username", username);
-        params.put("password", AESHelper.encryptAES(password, salt));
+        params.put("password", password);
 
 //        System.out.println("登陆参数 " + params);
 
         // 模拟登陆之前首先请求是否需要验证码接口
-        doc = Jsoup.connect(loginEntity.getNeedcaptchaUrl() + "?username=" + username)
-                .headers(headers)
-                .cookies(cookies)
-                .get();
-        boolean isNeedCaptcha = Boolean.valueOf(doc.body().text());
+        boolean isNeedCaptcha = true;
 //        System.out.println(isNeedCaptcha);
         if (isNeedCaptcha) {
             // 识别验证码后模拟登陆，最多尝试20次
             int time = TesseractOCRHelper.MAX_TRY_TIMES;
             while (time-- > 0) {
-                String code = ocrCaptcha(cookies, headers, loginEntity.getCaptchaUrl());
+                String code = ocrCaptcha(cookies, headers, loginEntity.getCaptchaUrl(), 4);
 //                System.out.println(code);
-                params.put("captchaResponse", code);
+                params.put("captcha", code);
+//                System.out.println("登陆参数 " + params);
                 Map<String, String> cookies2 = casSendLoginData(loginEntity.getLoginUrl(), cookies, params);
                 if (cookies2 != null) {
                     return cookies2;
@@ -149,17 +126,20 @@ public class WhpuCasLoginProcess {
      * @return
      * @throws Exception
      */
-    private Map<String, String> casSendLoginData(String login_url, Map<String, String> cookies, Map<String, String> params) throws Exception {
+    protected Map<String, String> casSendLoginData(String login_url, Map<String, String> cookies, Map<String, String> params) throws Exception {
+//        System.out.println(login_url);
         Map<String, String> headers = new HashMap<>();
         headers.put("Accept", "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9");
         headers.put("Accept-Encoding", "gzip, deflate");
-        headers.put("Accept-Language", "zh-CN,zh;q=0.9,en-US;q=0.8,en;q=0.7");
+        headers.put("Accept-Language", "zh-CN,zh;q=0.9");
         headers.put("Cache-Control", "max-age=0");
         headers.put("Connection", "keep-alive");
-        headers.put("Host", "cas.whpu.edu.cn");
-        headers.put("Referer", "http://cas.whpu.edu.cn/authserver/login?service=http%3A%2F%2Fehall.whpu.edu.cn%2Flogin%3Fservice%3Dhttp%3A%2F%2Fehall.whpu.edu.cn%2Fywtb-portal%2FLite%2Findex.html");
+        headers.put("Content-Type", "application/x-www-form-urlencoded");
+        headers.put("Host", "sso.cuit.edu.cn");
+        headers.put("Origin", "https://sso.cuit.edu.cn");
+        headers.put("Referer", "https://sso.cuit.edu.cn/authserver/login?service=http%3A%2F%2Fjwgl.cuit.edu.cn%2Feams%2Flogin.action");
         headers.put("Upgrade-Insecure-Requests", "1");
-        headers.put("User-Agent", "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/86.0.4240.111 Safari/537.36");
+        headers.put("User-Agent", "Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.125 Safari/537.36");
 
 //        System.out.println("headers " + headers);
 //        System.out.println("cookies " + cookies);
@@ -194,10 +174,10 @@ public class WhpuCasLoginProcess {
         } else if (login.statusCode() == HttpURLConnection.HTTP_OK) {
             // 登陆失败
             Document doc = login.parse();
-            System.out.println(doc);
-            Element msg = doc.getElementById("msg");
+//            System.out.println(doc);
+            Element msg = doc.getElementsByClass("textError").get(0);
 //            System.out.println(msg);
-            if (!msg.text().equals("验证码错误")) {
+            if (!msg.text().equals("验证码无效")) {
                 throw new RuntimeException(msg.text());
             }
         } else {
@@ -205,60 +185,5 @@ public class WhpuCasLoginProcess {
             throw new RuntimeException("教务系统服务器可能出错了，Http状态码是：" + login.statusCode());
         }
         return null;
-    }
-
-    /**
-     * 处理验证码识别
-     *
-     * @param cookies
-     * @param captcha_url
-     * @return
-     * @throws IOException
-     * @throws TesseractException
-     */
-    private String ocrCaptcha(Map<String, String> cookies, Map<String, String> headers, String captcha_url) throws IOException, TesseractException {
-        while (true) {
-            String filePach = System.getProperty("user.dir") + File.separator + System.currentTimeMillis() + ".jpg";
-//            System.out.println(filePach);
-//            System.out.println(captcha_url);
-            Connection.Response response = Jsoup.connect(captcha_url)
-                    .headers(headers)
-                    .cookies(cookies)
-                    .ignoreContentType(true)
-                    .execute();
-
-            // 四位验证码，背景有噪点
-            ImageHelper.saveImageFile(ImageHelper.binaryzation(response.bodyStream()), filePach);
-            String s = TesseractOCRHelper.doOcr(filePach);
-
-            File temp = new File(filePach);
-            temp.delete();
-
-            if (judge(s, 4)) {
-                return s;
-            }
-        }
-    }
-
-    /**
-     * 判断ocr识别出来的结果是否符合条件
-     *
-     * @param s
-     * @param len
-     * @return
-     */
-    private boolean judge(String s, int len) {
-        if (s == null || s.length() != len) {
-            return false;
-        }
-
-        for (int i = 0; i < s.length(); i++) {
-            char ch = s.charAt(i);
-            if (!(Character.isDigit(ch) || Character.isLetter(ch))) {
-                return false;
-            }
-        }
-
-        return true;
     }
 }
