+++ src/main/java/wiki/zimo/wiseduunifiedloginapi/process/impl/HenuCasLoginProcess.java
@@ -1,4 +1,4 @@
-package wiki.zimo.wiseduunifiedloginapi.process;
+package wiki.zimo.wiseduunifiedloginapi.process.impl;
 
 import net.sourceforge.tess4j.TesseractException;
 import org.jsoup.Connection;
@@ -8,31 +8,26 @@ import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 import wiki.zimo.wiseduunifiedloginapi.builder.CasLoginEntityBuilder;
 import wiki.zimo.wiseduunifiedloginapi.entity.CasLoginEntity;
+import wiki.zimo.wiseduunifiedloginapi.helper.AESHelper;
 import wiki.zimo.wiseduunifiedloginapi.helper.ImageHelper;
-import wiki.zimo.wiseduunifiedloginapi.helper.RSAHelper;
 import wiki.zimo.wiseduunifiedloginapi.helper.TesseractOCRHelper;
+import wiki.zimo.wiseduunifiedloginapi.process.OcrLoginProcess;
 import wiki.zimo.wiseduunifiedloginapi.trust.HttpsUrlValidator;
 
 import java.io.File;
 import java.io.IOException;
 import java.net.HttpURLConnection;
-import java.net.URL;
 import java.util.HashMap;
 import java.util.Map;
 
 /**
- * 安徽建筑大学认证
+ * 河南大学认证
  */
-public class AhjzuCasLoginProcess {
-    private CasLoginEntity loginEntity;
-    private Map<String, String> params;
-
-    public AhjzuCasLoginProcess(String loginUrl, Map<String, String> params) {
-        this.loginEntity = new CasLoginEntityBuilder()
-                .loginUrl(loginUrl)
-                .build();
-        loginEntity.setCaptchaUrl("http://sso.ahjzu.edu.cn/sso/code/code.jsp");
-        this.params = params;
+public class HenuCasLoginProcess extends OcrLoginProcess {
+    public HenuCasLoginProcess(String loginUrl, Map<String, String> params) {
+        super(loginUrl, params, CasLoginEntityBuilder.class);
+        loginEntity.setNeedcaptchaUrl("https://ids.henu.edu.cn/authserver/checkNeedCaptcha.htl");
+        loginEntity.setCaptchaUrl("https://ids.henu.edu.cn/authserver/getCaptcha.htl");
     }
 
     public Map<String, String> login() throws Exception {
@@ -40,31 +35,56 @@ public class AhjzuCasLoginProcess {
         // 忽略证书错误
         HttpsUrlValidator.retrieveResponseFromServer(loginEntity.getLoginUrl());
 
+        // 构造请求头
+        Map<String, String> headers = new HashMap<>();
+        headers.put("Accept", "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9");
+        headers.put("Accept-Encoding", "gzip, deflate");
+        headers.put("Accept-Language", "zh-CN,zh;q=0.9");
+        headers.put("Cache-Control", "max-age=0");
+        headers.put("Connection", "keep-alive");
+        headers.put("Host", "ids.henu.edu.cn");
+        headers.put("Upgrade-Insecure-Requests", "1");
+        headers.put("User-Agent", "Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.125 Safari/537.36");
+
+
         // 请求登陆页
         Connection con = Jsoup.connect(loginEntity.getLoginUrl())
-                .header("User-Agent", "Mozilla/5.0 (iPhone; CPU iPhone OS 13_2_3 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/13.0.3 Mobile/15E148 Safari/604.1")
-                .followRedirects(true);
+                .followRedirects(true)
+                .headers(headers);
         Connection.Response res = con.execute();
-
-        // 解析登陆页
-        Document doc = res.parse();
-        // 更新loginUrl
-        loginEntity.setLoginUrl(res.url().toString());
+//        System.out.println(res.url());
 //        System.out.println(doc);
 
         // 全局cookie
         Map<String, String> cookies = res.cookies();
 
+        // 解析登陆页
+        Document doc = Jsoup.connect(loginEntity.getLoginUrl())
+                .followRedirects(true)
+                .headers(headers)
+                .cookies(cookies)
+                .get();
+
         // 获取登陆表单
-        Element form = doc.getElementById("credentials");
-//        System.out.println(form);
+        Element pwdLoginDiv = doc.getElementById("pwdLoginDiv");
+        Element form = pwdLoginDiv.getElementById("loginFromId");
+
         if (form == null) {
-            throw new RuntimeException("网页中没有找到credentials，请联系开发者！！！");
+            throw new RuntimeException("网页中没有找到loginFromId，可能是教务系统已经更新，请联系开发者。");
         }
+
+        // 处理加密的盐
+        Element saltElement = doc.getElementById("pwdEncryptSalt");
+
+        String salt = null;
+        if (saltElement != null) {
+            salt = saltElement.val();
+        }
+
+//        System.out.println("盐是 " + salt);
 
         // 获取登陆表单里的输入
         Elements inputs = form.getElementsByTag("input");
-
 
         String username = this.params.get("username");
         String password = this.params.get("password");
@@ -72,17 +92,6 @@ public class AhjzuCasLoginProcess {
         // 构造post请求参数
         Map<String, String> params = new HashMap<>();
         for (Element e : inputs) {
-
-            // 填充用户名
-            if (e.attr("name").equals("username")) {
-                e.attr("value", username);
-            }
-
-            // 填充密码
-            if (e.attr("name").equals("password")) {
-                e.attr("value", RSAHelper.encrypt(password));
-            }
-
             // 排除空值表单属性
             if (e.attr("name").length() > 0) {
                 // 排除记住我
@@ -93,27 +102,25 @@ public class AhjzuCasLoginProcess {
             }
         }
 
+        params.put("username", username);
+        params.put("password", AESHelper.encryptAES(password, salt));
+
 //        System.out.println("登陆参数 " + params);
 
-        // 构造请求头
-        Map<String, String> headers = new HashMap<>();
-        headers.put("Accept", "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9");
-        headers.put("Accept-Encoding", "gzip, deflate");
-        headers.put("Cache-Control", "max-age=0");
-        headers.put("Connection", "keep-alive");
-        headers.put("Host", new URL(loginEntity.getLoginUrl()).getHost());
-        headers.put("Upgrade-Insecure-Requests", "1");
-        headers.put("User-Agent", "Mozilla/5.0 (iPhone; CPU iPhone OS 13_2_3 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/13.0.3 Mobile/15E148 Safari/604.1");
-
-        boolean isNeedCaptcha = true;
+        // 模拟登陆之前首先请求是否需要验证码接口
+        doc = Jsoup.connect(loginEntity.getNeedcaptchaUrl() + "?username=" + username)
+                .headers(headers)
+                .cookies(cookies)
+                .get();
+        boolean isNeedCaptcha = Boolean.valueOf(doc.body().text());
 //        System.out.println(isNeedCaptcha);
         if (isNeedCaptcha) {
             // 识别验证码后模拟登陆，最多尝试20次
             int time = TesseractOCRHelper.MAX_TRY_TIMES;
             while (time-- > 0) {
-                String code = ocrCaptcha(cookies, headers, loginEntity.getCaptchaUrl());
+                String code = ocrCaptcha(cookies, headers, loginEntity.getCaptchaUrl(), 4);
 //                System.out.println(code);
-                params.put("validateCode", code);
+                params.put("captchaResponse", code);
                 Map<String, String> cookies2 = casSendLoginData(loginEntity.getLoginUrl(), cookies, params);
                 if (cookies2 != null) {
                     return cookies2;
@@ -136,10 +143,26 @@ public class AhjzuCasLoginProcess {
      * @return
      * @throws Exception
      */
-    private Map<String, String> casSendLoginData(String login_url, Map<String, String> cookies, Map<String, String> params) throws Exception {
+    protected Map<String, String> casSendLoginData(String login_url, Map<String, String> cookies, Map<String, String> params) throws Exception {
+        Map<String, String> headers = new HashMap<>();
+        headers.put("Accept", "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9");
+        headers.put("Accept-Encoding", "gzip, deflate");
+        headers.put("Accept-Language", "zh-CN,zh;q=0.9");
+        headers.put("Cache-Control", "max-age=0");
+        headers.put("Connection", "keep-alive");
+        headers.put("Content-Type", "application/x-www-form-urlencoded");
+        headers.put("Host", "ids.henu.edu.cn");
+        headers.put("Origin", "https://ids.henu.edu.cn");
+        headers.put("Referer", "https://ids.henu.edu.cn/authserver/login?service=https%3A%2F%2Fhenu.campusphere.net%2Fportal%2Flogin");
+        headers.put("Upgrade-Insecure-Requests", "1");
+        headers.put("User-Agent", "Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.125 Safari/537.36");
+
+//        System.out.println("headers " + headers);
+//        System.out.println("cookies " + cookies);
+
         Connection con = Jsoup.connect(login_url);
 //        System.out.println(login_url);
-        Connection.Response login = con.ignoreContentType(true)
+        Connection.Response login = con.headers(headers)
                 .followRedirects(false)
                 .method(Connection.Method.POST)
                 .data(params)
@@ -167,9 +190,10 @@ public class AhjzuCasLoginProcess {
         } else if (login.statusCode() == HttpURLConnection.HTTP_OK) {
             // 登陆失败
             Document doc = login.parse();
-            Element msg = doc.getElementsByClass("err").get(0);
+            System.out.println(doc);
+            Element msg = doc.getElementById("msg");
 //            System.out.println(msg);
-            if (!msg.text().equals("验证码不正确")) {
+            if (!msg.text().equals("验证码错误")) {
                 throw new RuntimeException(msg.text());
             }
         } else {
@@ -178,58 +202,5 @@ public class AhjzuCasLoginProcess {
         }
         return null;
     }
-
-    /**
-     * 处理验证码识别
-     *
-     * @param cookies
-     * @param captcha_url
-     * @return
-     * @throws IOException
-     * @throws TesseractException
-     */
-    private String ocrCaptcha(Map<String, String> cookies, Map<String, String> headers, String captcha_url) throws IOException, TesseractException {
-        while (true) {
-            String filePach = System.getProperty("user.dir") + File.separator + System.currentTimeMillis() + ".jpg";
-//            System.out.println(filePach);
-//            System.out.println(captcha_url);
-            Connection.Response response = Jsoup.connect(captcha_url)
-                    .headers(headers).cookies(cookies)
-                    .ignoreContentType(true)
-                    .execute();
-
-            // 四位验证码，背景有噪点
-            ImageHelper.saveImageFile(ImageHelper.binaryzation(response.bodyStream()), filePach);
-            String s = TesseractOCRHelper.doOcr(filePach);
-
-            File temp = new File(filePach);
-            temp.delete();
-
-            if (judge(s, 4)) {
-                return s;
-            }
-        }
-    }
-
-    /**
-     * 判断ocr识别出来的结果是否符合条件
-     *
-     * @param s
-     * @param len
-     * @return
-     */
-    private boolean judge(String s, int len) {
-        if (s == null || s.length() != len) {
-            return false;
-        }
-
-        for (int i = 0; i < s.length(); i++) {
-            char ch = s.charAt(i);
-            if (!(Character.isDigit(ch) || Character.isLetter(ch))) {
-                return false;
-            }
-        }
-
-        return true;
-    }
 }
+
