+++ src/main/java/wiki/zimo/wiseduunifiedloginapi/process/impl/HytcCasLoginProcess.java
@@ -1,6 +1,5 @@
-package wiki.zimo.wiseduunifiedloginapi.process;
+package wiki.zimo.wiseduunifiedloginapi.process.impl;
 
-import net.sourceforge.tess4j.TesseractException;
 import org.jsoup.Connection;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
@@ -8,30 +7,21 @@ import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 import wiki.zimo.wiseduunifiedloginapi.builder.CasLoginEntityBuilder;
 import wiki.zimo.wiseduunifiedloginapi.entity.CasLoginEntity;
-import wiki.zimo.wiseduunifiedloginapi.helper.ImageHelper;
 import wiki.zimo.wiseduunifiedloginapi.helper.RSAHelper;
-import wiki.zimo.wiseduunifiedloginapi.helper.TesseractOCRHelper;
+import wiki.zimo.wiseduunifiedloginapi.process.OcrLoginProcess;
 import wiki.zimo.wiseduunifiedloginapi.trust.HttpsUrlValidator;
 
-import java.io.File;
-import java.io.IOException;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.util.HashMap;
 import java.util.Map;
 
 /**
- * 昆明学院认证
+ * 淮阴师范学院认证
  */
-public class KmuCasLoginProcess {
-    private CasLoginEntity loginEntity;
-    private Map<String, String> params;
-
-    public KmuCasLoginProcess(String loginUrl, Map<String, String> params) {
-        this.loginEntity = new CasLoginEntityBuilder()
-                .loginUrl(loginUrl)
-                .build();
-        this.params = params;
+public class HytcCasLoginProcess extends OcrLoginProcess {
+    public HytcCasLoginProcess(String loginUrl, Map<String, String> params) {
+        super(loginUrl, params, CasLoginEntityBuilder.class);
     }
 
     public Map<String, String> login() throws Exception {
@@ -45,13 +35,10 @@ public class KmuCasLoginProcess {
                 .followRedirects(true);
         Connection.Response res = con.execute();
 
-        // 更新loginUrl
         loginEntity.setLoginUrl(res.url().toString());
-        loginEntity.setCaptchaUrl("http://cas.kmu.edu.cn/lyuapServer/captcha.jsp");
 
         // 解析登陆页
         Document doc = res.parse();
-
 //        System.out.println(doc);
 
         // 全局cookie
@@ -82,7 +69,7 @@ public class KmuCasLoginProcess {
 
             // 填充密码
             if (e.attr("name").equals("password")) {
-                e.attr("value", RSAHelper.encrypt2(password));
+                e.attr("value", RSAHelper.encrypt4(password));
             }
 
             // 排除空值表单属性
@@ -107,27 +94,8 @@ public class KmuCasLoginProcess {
         headers.put("Upgrade-Insecure-Requests", "1");
         headers.put("User-Agent", "Mozilla/5.0 (iPhone; CPU iPhone OS 13_2_3 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/13.0.3 Mobile/15E148 Safari/604.1");
 
-        // 模拟登陆之前首先请求是否需要验证码接口
-        boolean isNeedCaptcha = true;
-//        System.out.println(isNeedCaptcha);
-        if (isNeedCaptcha) {
-            // 识别验证码后模拟登陆，最多尝试20次
-            int time = TesseractOCRHelper.MAX_TRY_TIMES;
-            while (time-- > 0) {
-                String code = ocrCaptcha(cookies, headers, loginEntity.getCaptchaUrl());
-//                System.out.println(code);
-                params.put("captcha", code);
-                Map<String, String> cookies2 = casSendLoginData(loginEntity.getLoginUrl(), cookies, params);
-                if (cookies2 != null) {
-                    return cookies2;
-                }
-            }
-            // 执行到这里就代表验证码识别尝试已经达到了最大的次数
-            throw new RuntimeException("验证码识别错误，请重试");
-        } else {
-            // 直接模拟登陆
-            return casSendLoginData(loginEntity.getLoginUrl(), cookies, params);
-        }
+        // 不需要验证码过程，或者说验证码过程有设计缺陷，根本不会触发，所以直接模拟登陆
+        return casSendLoginData(loginEntity.getLoginUrl(), cookies, params);
     }
 
     /**
@@ -139,7 +107,7 @@ public class KmuCasLoginProcess {
      * @return
      * @throws Exception
      */
-    private Map<String, String> casSendLoginData(String login_url, Map<String, String> cookies, Map<String, String> params) throws Exception {
+    protected Map<String, String> casSendLoginData(String login_url, Map<String, String> cookies, Map<String, String> params) throws Exception {
         Connection con = Jsoup.connect(login_url);
 //        System.out.println(login_url);
         Connection.Response login = con.ignoreContentType(true).followRedirects(false).method(Connection.Method.POST).data(params).cookies(cookies).execute();
@@ -156,6 +124,7 @@ public class KmuCasLoginProcess {
                     .ignoreContentType(true)
                     .followRedirects(true)
                     .method(Connection.Method.POST)
+                    .header("User-Agent", "Mozilla/5.0 (iPhone; CPU iPhone OS 13_2_3 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/13.0.3 Mobile/15E148 Safari/604.1")
                     .cookies(cookies);
             // 请求，再次更新cookie
             login = con.execute();
@@ -167,67 +136,10 @@ public class KmuCasLoginProcess {
             Document doc = login.parse();
             Element msg = doc.getElementById("msg");
 //            System.out.println(msg);
-            if (!msg.text().equals("验证码不正确")) {
-                throw new RuntimeException(msg.text());
-            }
+            throw new RuntimeException(msg.text());
         } else {
             // 服务器可能出错
             throw new RuntimeException("教务系统服务器可能出错了，Http状态码是：" + login.statusCode());
         }
-        return null;
-    }
-
-    /**
-     * 处理验证码识别
-     *
-     * @param cookies
-     * @param captcha_url
-     * @return
-     * @throws IOException
-     * @throws TesseractException
-     */
-    private String ocrCaptcha(Map<String, String> cookies, Map<String, String> headers, String captcha_url) throws IOException, TesseractException {
-        while (true) {
-            String filePach = System.getProperty("user.dir") + File.separator + System.currentTimeMillis() + ".jpg";
-//            System.out.println(filePach);
-//            System.out.println(captcha_url);
-            Connection.Response response = Jsoup.connect(captcha_url)
-                    .headers(headers).cookies(cookies)
-                    .ignoreContentType(true)
-                    .execute();
-
-            // 四位验证码，背景有噪点
-            ImageHelper.saveImageFile(ImageHelper.binaryzation(response.bodyStream()), filePach);
-            String s = TesseractOCRHelper.doOcr(filePach);
-
-            File temp = new File(filePach);
-            temp.delete();
-
-            if (judge(s, 4)) {
-                return s;
-            }
-        }
-    }
-
-    /**
-     * 判断ocr识别出来的结果是否符合条件
-     *
-     * @param s
-     * @param len
-     * @return
-     */
-    private boolean judge(String s, int len) {
-        if (s == null || s.length() != len) {
-            return false;
-        }
-
-        for (int i = 0; i < s.length(); i++) {
-            char ch = s.charAt(i);
-            if (!(Character.isDigit(ch) || Character.isLetter(ch))) {
-                return false;
-            }
-        }
-
-        return true;
     }
 }
