+++ src/main/java/wiki/zimo/wiseduunifiedloginapi/process/impl/WhpuCasLoginProcess.java
@@ -1,4 +1,4 @@
-package wiki.zimo.wiseduunifiedloginapi.process;
+package wiki.zimo.wiseduunifiedloginapi.process.impl;
 
 import net.sourceforge.tess4j.TesseractException;
 import org.jsoup.Connection;
@@ -11,6 +11,7 @@ import wiki.zimo.wiseduunifiedloginapi.entity.CasLoginEntity;
 import wiki.zimo.wiseduunifiedloginapi.helper.AESHelper;
 import wiki.zimo.wiseduunifiedloginapi.helper.ImageHelper;
 import wiki.zimo.wiseduunifiedloginapi.helper.TesseractOCRHelper;
+import wiki.zimo.wiseduunifiedloginapi.process.OcrLoginProcess;
 import wiki.zimo.wiseduunifiedloginapi.trust.HttpsUrlValidator;
 
 import java.io.File;
@@ -20,19 +21,14 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
- * 河南大学认证
+ * 武汉轻工大学认证
  */
-public class HenuCasLoginProcess {
-    private CasLoginEntity loginEntity;
-    private Map<String, String> params;
+public class WhpuCasLoginProcess extends OcrLoginProcess {
 
-    public HenuCasLoginProcess(String loginUrl, Map<String, String> params) {
-        this.loginEntity = new CasLoginEntityBuilder()
-                .loginUrl(loginUrl)
-                .build();
-        loginEntity.setNeedcaptchaUrl("https://ids.henu.edu.cn/authserver/checkNeedCaptcha.htl");
-        loginEntity.setCaptchaUrl("https://ids.henu.edu.cn/authserver/getCaptcha.htl");
-        this.params = params;
+    public WhpuCasLoginProcess(String loginUrl, Map<String, String> params) {
+        super(loginUrl, params, CasLoginEntityBuilder.class);
+        loginEntity.setNeedcaptchaUrl("http://cas.whpu.edu.cn/authserver/checkNeedCaptcha.htl");
+        loginEntity.setCaptchaUrl("http://cas.whpu.edu.cn/authserver/getCaptcha.htl");
     }
 
     public Map<String, String> login() throws Exception {
@@ -47,7 +43,7 @@ public class HenuCasLoginProcess {
         headers.put("Accept-Language", "zh-CN,zh;q=0.9");
         headers.put("Cache-Control", "max-age=0");
         headers.put("Connection", "keep-alive");
-        headers.put("Host", "ids.henu.edu.cn");
+        headers.put("Host", "cas.whpu.edu.cn");
         headers.put("Upgrade-Insecure-Requests", "1");
         headers.put("User-Agent", "Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.125 Safari/537.36");
 
@@ -75,7 +71,7 @@ public class HenuCasLoginProcess {
         Element form = pwdLoginDiv.getElementById("loginFromId");
 
         if (form == null) {
-            throw new RuntimeException("网页中没有找到loginFromId，可能是教务系统已经更新，请联系开发者。");
+            throw new RuntimeException("网页中没有找到loginFromId，请联系开发者！！！");
         }
 
         // 处理加密的盐
@@ -90,6 +86,7 @@ public class HenuCasLoginProcess {
 
         // 获取登陆表单里的输入
         Elements inputs = form.getElementsByTag("input");
+
 
         String username = this.params.get("username");
         String password = this.params.get("password");
@@ -123,7 +120,7 @@ public class HenuCasLoginProcess {
             // 识别验证码后模拟登陆，最多尝试20次
             int time = TesseractOCRHelper.MAX_TRY_TIMES;
             while (time-- > 0) {
-                String code = ocrCaptcha(cookies, headers, loginEntity.getCaptchaUrl());
+                String code = ocrCaptcha(cookies, headers, loginEntity.getCaptchaUrl(), 4);
 //                System.out.println(code);
                 params.put("captchaResponse", code);
                 Map<String, String> cookies2 = casSendLoginData(loginEntity.getLoginUrl(), cookies, params);
@@ -148,19 +145,17 @@ public class HenuCasLoginProcess {
      * @return
      * @throws Exception
      */
-    private Map<String, String> casSendLoginData(String login_url, Map<String, String> cookies, Map<String, String> params) throws Exception {
+    protected Map<String, String> casSendLoginData(String login_url, Map<String, String> cookies, Map<String, String> params) throws Exception {
         Map<String, String> headers = new HashMap<>();
         headers.put("Accept", "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9");
         headers.put("Accept-Encoding", "gzip, deflate");
-        headers.put("Accept-Language", "zh-CN,zh;q=0.9");
+        headers.put("Accept-Language", "zh-CN,zh;q=0.9,en-US;q=0.8,en;q=0.7");
         headers.put("Cache-Control", "max-age=0");
         headers.put("Connection", "keep-alive");
-        headers.put("Content-Type", "application/x-www-form-urlencoded");
-        headers.put("Host", "ids.henu.edu.cn");
-        headers.put("Origin", "https://ids.henu.edu.cn");
-        headers.put("Referer", "https://ids.henu.edu.cn/authserver/login?service=https%3A%2F%2Fhenu.campusphere.net%2Fportal%2Flogin");
+        headers.put("Host", "cas.whpu.edu.cn");
+        headers.put("Referer", "http://cas.whpu.edu.cn/authserver/login?service=http%3A%2F%2Fehall.whpu.edu.cn%2Flogin%3Fservice%3Dhttp%3A%2F%2Fehall.whpu.edu.cn%2Fywtb-portal%2FLite%2Findex.html");
         headers.put("Upgrade-Insecure-Requests", "1");
-        headers.put("User-Agent", "Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.125 Safari/537.36");
+        headers.put("User-Agent", "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/86.0.4240.111 Safari/537.36");
 
 //        System.out.println("headers " + headers);
 //        System.out.println("cookies " + cookies);
@@ -207,60 +202,4 @@ public class HenuCasLoginProcess {
         }
         return null;
     }
-
-    /**
-     * 处理验证码识别
-     *
-     * @param cookies
-     * @param captcha_url
-     * @return
-     * @throws IOException
-     * @throws TesseractException
-     */
-    private String ocrCaptcha(Map<String, String> cookies, Map<String, String> headers, String captcha_url) throws IOException, TesseractException {
-        while (true) {
-            String filePach = System.getProperty("user.dir") + File.separator + System.currentTimeMillis() + ".jpg";
-//            System.out.println(filePach);
-//            System.out.println(captcha_url);
-            Connection.Response response = Jsoup.connect(captcha_url)
-                    .headers(headers)
-                    .cookies(cookies)
-                    .ignoreContentType(true)
-                    .execute();
-
-            // 四位验证码，背景有噪点
-            ImageHelper.saveImageFile(ImageHelper.binaryzation(response.bodyStream()), filePach);
-            String s = TesseractOCRHelper.doOcr(filePach);
-
-            File temp = new File(filePach);
-            temp.delete();
-
-            if (judge(s, 4)) {
-                return s;
-            }
-        }
-    }
-
-    /**
-     * 判断ocr识别出来的结果是否符合条件
-     *
-     * @param s
-     * @param len
-     * @return
-     */
-    private boolean judge(String s, int len) {
-        if (s == null || s.length() != len) {
-            return false;
-        }
-
-        for (int i = 0; i < s.length(); i++) {
-            char ch = s.charAt(i);
-            if (!(Character.isDigit(ch) || Character.isLetter(ch))) {
-                return false;
-            }
-        }
-
-        return true;
-    }
 }
-
